fix: change return type of sum, squareSum and relevant variable types.

+++ core/ImgStatics.hpp
@@ -23,7 +23,7 @@ class ImgStatics {
      * @param src MatrixBuffer<BYTE> to calculate
      * @return sum of elements
      */
-    float sum(MatrixBuffer<BYTE> &src);
+    long long sum(MatrixBuffer<BYTE> &src);
 
     /**
      * @brief Get average of elements in buffer.
@@ -32,7 +32,7 @@ class ImgStatics {
      * @param src MatrixBuffer<BYTE> to calculate
      * @return Average of elements
      */
-    float mean(MatrixBuffer<BYTE> &src);
+    double mean(MatrixBuffer<BYTE> &src);
 
     /**
      * @brief Get variance of elements in buffer.
@@ -41,7 +41,7 @@ class ImgStatics {
      * @param src MatrixBuffer<BYTE> to calculate
      * @return Variance of elements
      */
-    float var(MatrixBuffer<BYTE> &src);
+    double var(MatrixBuffer<BYTE> &src);
 
     /**
      * @brief Get Sum of x^2 in buffer.
@@ -50,5 +50,5 @@ class ImgStatics {
      * @param src MatrixBuffer<BYTE> to calculate
      * @return Sum of x^2 in elements
      */
-    float squareSum(MatrixBuffer<BYTE> &src);
+    long long squareSum(MatrixBuffer<BYTE> &src);
 };
